refactor: move non-route code out of app/
app/ is the routing tree, but it also held components, project data and
stylesheets. None of those are routes, and Next's convention is to keep
them outside the router.

  src/app/components  -> src/components
  src/app/assets      -> src/data
  src/app/*.scss      -> src/styles

Styles had to move too: once components live outside app/, their variable
imports would otherwise have to reach back into the routing tree.

Imports now use the @/* alias that already existed in tsconfig but was
unused. Also drops the unreferenced create-next-app svgs and points
sassOptions.includePaths at src/styles, which previously named a directory
that did not exist.
