+++ src/app/contact/contact.module.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../styles/variables.scss';
 
 // Local to this component. Promote it to variables.scss if anything else
 // ever needs a non-text white.
