+++ src/app/page.module.scss
@@ -1,4 +1,4 @@
-@import 'variables.scss';
+@import '../styles/variables.scss';
 
 .hero {
     height: 100vh;
