+++ src/components/nav/nav.module.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../styles/variables.scss';
 .nav{
     display: flex;
     justify-content: space-between;
