+++ src/components/project-preview/project-preview.module.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../styles/variables.scss';
 
 .project{
     height: 100vh;
